This pull request integrates many bugfixes and new features.
Among many other things:

- Fixed several memory issues
- New Icons
- Build process for production and development
- Undo/Reso bugfixes
- Cmake target rwa trace for testing

+++ rwastateview.cpp
@@ -6,13 +6,18 @@ RwaStateView::RwaStateView(QWidget* parent, RwaScene *scene, QString name)
 {
     setAcceptDrops(true);
     setAlignment(Qt::AlignTop);
+
     windowSplitter = new QSplitter(this);
     connect(windowSplitter, SIGNAL(splitterMoved(int,int)), this, SLOT(handleSplitter(int, int)));
+
     layout = new QBoxLayout(QBoxLayout::LeftToRight,this);
     layout->setContentsMargins(0,0,0,0);
+
     assetList = new RwaAssetList(this, scene);
     assetAttributes = new RwaAssetAttributeView(this, scene);
-    assetAttributes->scrollArea->setMaximumWidth(250);
+    assetAttributes->scrollArea->setFixedWidth(260);
+    assetAttributes->scrollArea->setHorizontalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
+
     currentPoint = nullptr;
     editStateRadius = false;
     editStateHeight = false;
@@ -22,8 +27,9 @@ RwaStateView::RwaStateView(QWidget* parent, RwaScene *scene, QString name)
     assetReflectionsVisible = true;
     assetsVisible = true;
     onlyAssetsOfCurrentStateVisible = true;
+    onlyRadiiOfCurrentStateVisible = true;
 
-    connect(backend, SIGNAL(sendMoveCurrentState1(double, double)),
+    connect(backend, SIGNAL(sendMovePixmapsOfCurrentState1(double, double)),
               this, SLOT(movePixmapsOfCurrentState(double,double)));
 
     connect(backend, SIGNAL(sendMoveCurrentScene()),
@@ -32,10 +38,10 @@ RwaStateView::RwaStateView(QWidget* parent, RwaScene *scene, QString name)
     connect(this, SIGNAL(sendMoveCurrentState1(double, double)),
               backend, SLOT(receiveMoveCurrentState1(double,double)));
 
-    connect(backend, SIGNAL(sendMoveCurrentAsset1(double, double)),
+    connect(backend, SIGNAL(sendMovePixmapsOfCurrentAsset1(double, double)),
               this, SLOT(movePixmapsOfCurrentAsset(double,double)));
 
-    connect(this, SIGNAL(sendMoveCurrentAsset1(double, double)),
+    connect(this, SIGNAL(sendMovePixmapsOfCurrentAsset1(double, double)),
               backend, SLOT(receiveMoveCurrentAsset1(double,double)));
 
     connect(this, SIGNAL(sendMoveCurrentAssetChannel(double, double, int)),
@@ -74,7 +80,11 @@ RwaStateView::RwaStateView(QWidget* parent, RwaScene *scene, QString name)
     connect(backend, SIGNAL(sendCurrentStateRadiusEdited()),
               this, SLOT(receiveUpdateCurrentStateRadius()));
 
+    connect(backend, SIGNAL(sendCurrentStateWithoutRepositioning(RwaState *)),
+              this, SLOT(setCurrentStateWithoutRepositioning(RwaState *)));
+
     addZoomButtons();
+
     windowSplitter->addWidget(assetAttributes->scrollArea);
     windowSplitter->addWidget(assetList);
     windowSplitter->addWidget(mc);
@@ -92,24 +102,24 @@ void RwaStateView::keyPressEvent(QKeyEvent *event)
     {
         case Qt::Key_Return:
         case Qt::Key_Enter:
-          //qDebug() << "Enter";
-          break;
+            // qDebug() << "Enter";
+            break;
         case Qt::Key_Escape:
-          //qDebug() << "Escape";
-          break;
+            // qDebug() << "Escape";
+            break;
         case Qt::Key_Insert:
-          //qDebug() << "Insert";
-          break;
+            // qDebug() << "Insert";
+            break;
 
-          default:
-        //qDebug() << "RwaStateView::keyPressEvent" << event->key();
-          break;
+        default:
+            // qDebug() << "RwaStateView::keyPressEvent" << event->key();
+            break;
      }
 }
 
 void RwaStateView::correctMapView()
 {
-     mc->resize(QSize(windowSplitter->sizes().at(2), this->height()));
+    mc->resize(QSize(windowSplitter->sizes().at(2), this->height()));
 }
 
 void RwaStateView::receiveUpdateCurrentStateRadius()
@@ -156,14 +166,6 @@ void RwaStateView::setCurrentAsset(RwaAsset1 *asset)
         return;
 
     RwaGraphicsView::setCurrentAsset(asset);
-
-//    currentAsset = asset;
-//    currentState->setLastTouchedAsset(this->currentAsset);
-//    updateAssetPixmaps();
-//    updateReflectionPixmaps();
-//    if(!(QObject::sender() == backend))
-//        emit sendCurrentAsset(currentAsset);
-
 }
 
 void RwaStateView::redrawStateRadii()
@@ -189,18 +191,16 @@ void RwaStateView::setMapCoordinates(QPointF coordinates)
 
 void RwaStateView::adaptSize(qint32 width, qint32 height)
 {
+    (void) width;
+    (void) height;
     mc->resize(QSize(windowSplitter->sizes().at(2), this->height()));
-//    QSettings settings("Intrinsic Audio", "Rwa Creator");
-//    settings.setValue("stateViewGeometry", windowSplitter->saveGeometry());
-//    settings.setValue("stateViewState", windowSplitter->saveState());
 }
 
 void RwaStateView::handleSplitter(int pos, int index)
 {
-     mc->resize(QSize(windowSplitter->sizes().at(2), this->height()));
-//    QSettings settings("Intrinsic Audio", "Rwa Creator");
-//    settings.setValue("stateViewGeometry", windowSplitter->saveGeometry());
-//    settings.setValue("stateViewState", windowSplitter->saveState());
+    (void) pos;
+    (void) index;
+    mc->resize(QSize(windowSplitter->sizes().at(2), this->height()));
 }
 
 void RwaStateView::receiveMouseMoveEvent(const QMouseEvent*, const QPointF myPoint)
@@ -208,7 +208,7 @@ void RwaStateView::receiveMouseMoveEvent(const QMouseEvent*, const QPointF myPoi
     if(!currentState)
         return;
 
-    RwaMapItem *geo = (RwaMapItem *)currentMapItem;
+    RwaMapItem *geo = static_cast<RwaMapItem *>(currentMapItem);
     QPointF lastCoordinate;
     double dx, dy;
     std::vector<double> tmp(2, 0.0);
@@ -220,79 +220,57 @@ void RwaStateView::receiveMouseMoveEvent(const QMouseEvent*, const QPointF myPoi
     if(backend->isSimulationRunning())
         return;
 
-    if(editStatePosition)
-    {
-        lastCoordinate = QPointF(currentState->getCoordinates()[0], currentState->getCoordinates()[1]) ;
-        tmp[0] = mc->currentCoordinate().x();
-        tmp[1] = mc->currentCoordinate().y();
-        currentState->setCoordinates(tmp);
-        dx = currentState->getCoordinates()[0] - lastCoordinate.x();
-        dy = currentState->getCoordinates()[1] - lastCoordinate.y();
-        if(currentState->childrenDoFollowMe())
-            currentState->moveMyChildren(dx, dy);
-
-        emit sendMoveCurrentState1(dx, dy);
-        setUndoAction("Move State");
-
-        return;
-    }
-
     if(geo)
     {
         if(currentAsset->getLockPosition())
             return;
 
-         if(geo->getRwaType() == RWAPOSITIONTYPE_ASSET)
-         {
+        tmp = {myPoint.x(), myPoint.y()};
+
+        if(geo->getRwaType() == RWAPOSITIONTYPE_ASSET)
+        {
             lastCoordinate = QPointF(currentAsset->getCoordinates()[0], currentAsset->getCoordinates()[1]);
-            tmp[0] = myPoint.x();
-            tmp[1] = myPoint.y();
-            currentAsset->setCoordinates(tmp);
             dx = myPoint.x() - lastCoordinate.x();
             dy = myPoint.y() - lastCoordinate.y();
+            currentAsset->setCoordinates(tmp);
             currentAsset->moveMyChildren(dx, dy);
-            emit sendMoveCurrentAsset1(dx, dy);
+            emit sendMovePixmapsOfCurrentAsset1(dx, dy);
             setUndoAction("Move Asset");
-         }
+        }
 
-         if(geo->getRwaType() == RWAPOSITIONTYPE_ASSETSTARTPOINT)
-         {
-             tmp[0] = myPoint.x();
-             tmp[1] = myPoint.y();
-             currentAsset->setStartPosition(tmp);
-             geo->setCoordinate(myPoint);
-             setUndoAction("Move Asset start location.");
-         }
+        if(geo->getRwaType() == RWAPOSITIONTYPE_ASSETSTARTPOINT)
+        {
+            currentAsset->setStartPosition(tmp);
+            geo->setCoordinate(myPoint);
+            setUndoAction("Move Asset start location.");
+        }
 
-         if(geo->getRwaType() == RWAPOSITIONTYPE_ASSETCHANNEL &&
-                 currentAsset->individuellChannelPositionsAllowed() )
-         {
-             lastCoordinate = QPointF(currentAsset->channelcoordinates[geo->getChannel()][0], currentAsset->channelcoordinates[geo->getChannel()][1]);
-             dx = myPoint.x() - lastCoordinate.x();
-             dy = myPoint.y() - lastCoordinate.y();
-             tmp[0] = myPoint.x();
-             tmp[1] = myPoint.y();
-             currentAsset->setChannelCoordinate(geo->getChannel(), tmp);
-             currentAsset->individuellChannelPosition[geo->getChannel()] = true;
-             geo->setCoordinate(myPoint);
-             emit sendMoveCurrentAssetChannel(dx, dy, geo->getChannel());
-             setUndoAction("Move Asset channel position.");
-         }
+        if(geo->getRwaType() == RWAPOSITIONTYPE_ASSETCHANNEL &&
+             currentAsset->customChannelPositionsEnabled() )
+        {
+            lastCoordinate = QPointF(currentAsset->channelcoordinates[geo->getChannel()][0], currentAsset->channelcoordinates[geo->getChannel()][1]);
+            dx = myPoint.x() - lastCoordinate.x();
+            dy = myPoint.y() - lastCoordinate.y();
+            currentAsset->setChannelCoordinate(geo->getChannel(), tmp);
 
-         if(geo->getRwaType() == RWAPOSITIONTYPE_REFLECTIONPOSITION)
-         {
-             lastCoordinate = QPointF(currentAsset->reflectioncoordinates[geo->getChannel()][0], currentAsset->reflectioncoordinates[geo->getChannel()][1]);
-             dx = myPoint.x() - lastCoordinate.x();
-             dy = myPoint.y() - lastCoordinate.y();
-             tmp[0] = myPoint.x();
-             tmp[1] = myPoint.y();
-             currentAsset->setReflectionCoordinate(geo->getChannel(), tmp);
-             geo->setCoordinate(myPoint);
-             emit sendMoveCurrentAssetReflection(dx, dy, geo->getChannel());
-             setUndoAction("Move Asset reflection position.");
-         }
+            currentAsset->hasCustomChannelPosition[geo->getChannel()] = true;
+            geo->setCoordinate(myPoint);
+            emit sendMoveCurrentAssetChannel(dx, dy, geo->getChannel());
+            setUndoAction("Move Asset channel position.");
+        }
 
-         return;
+        if(geo->getRwaType() == RWAPOSITIONTYPE_REFLECTIONPOSITION)
+        {
+            lastCoordinate = QPointF(currentAsset->reflectioncoordinates[geo->getChannel()][0], currentAsset->reflectioncoordinates[geo->getChannel()][1]);
+            dx = myPoint.x() - lastCoordinate.x();
+            dy = myPoint.y() - lastCoordinate.y();
+            currentAsset->setReflectionCoordinate(geo->getChannel(), tmp);
+            geo->setCoordinate(myPoint);
+            emit sendMoveCurrentAssetReflection(dx, dy, geo->getChannel());
+            setUndoAction("Move Asset reflection position.");
+        }
+
+        return;
     }
 
     if(editArea)
@@ -352,7 +330,7 @@ void RwaStateView::receiveMouseDownEvent(const QMouseEvent *event, const QPointF
                  }
                  else
                  {
-                     if(asset->individuellChannelPositionsAllowed())
+                     if(asset->customChannelPositionsEnabled())
                      {
                          emit sendCurrentAsset(asset);
                          mc->setMouseMode(MapControl::None);
@@ -375,9 +353,6 @@ void RwaStateView::receiveMouseDownEvent(const QMouseEvent *event, const QPointF
              }
          }
      }
-
-     if(!currentState->positionIsLocked())
-        editStatePosition = true;
 }
 
 void RwaStateView::receiveMouseReleaseEvent()
@@ -402,6 +377,7 @@ void RwaStateView::receiveMouseReleaseEvent()
 
 void RwaStateView::dropEvent(QDropEvent *event)
 {
+    (void) event;
     if(backend->logOther)
         qDebug();
 }
@@ -425,14 +401,13 @@ void RwaStateView::setCurrentState(RwaState *state)
     window->setWindowTitle("State View - "+ QString::fromStdString(state->objectName()));
     mc->setView(QPointF(state->getCoordinates()[0], state->getCoordinates()[1]));
     setMap2AreaZoomLevel(state);
-    RwaGraphicsView::setCurrentState(state);    
+    RwaGraphicsView::setCurrentState(state);
 
     if(QObject::sender() != this->backend)
     {
         qDebug();
         emit sendCurrentAsset(state->lastTouchedAsset);
     }
-
 }
 
 void RwaStateView::setCurrentScene(RwaScene *scene)
@@ -452,6 +427,19 @@ void RwaStateView::setCurrentScene(RwaScene *scene)
     }
 }
 
+void RwaStateView::setCurrentStateWithoutRepositioning(RwaState *state)
+{
+    if(!state)
+        return;
+
+    QDockWidget *window = static_cast<QDockWidget *>(parent());
+    window->setWindowTitle("State View - "+ QString::fromStdString(state->objectName()));
+    RwaGraphicsView::setCurrentState(state);
+
+    if(QObject::sender() != this->backend)
+        emit sendCurrentAsset(state->lastTouchedAsset);
+}
+
 int RwaStateView::getNumberOfSelectedAssets()
 {
     return assetList->getNumberOfSelectedAssets();
@@ -463,35 +451,43 @@ void RwaStateView::addAssetItem(const QString &path, qint32 type)
     {
         string uid = std::string(QUuid::createUuid().toString().toLatin1());
         RwaAsset1 *newItem = new RwaAsset1(path.toStdString(), currentState->getCoordinates(), type, uid);
-        backend->simulator->runtime->openFile4MetaData(path.toLatin1());
-        newItem->setDuration(((RwaRuntime::assetDuration/RwaRuntime::assetSampleRate)*1000.0));
-        newItem->setNumberOfChannels(RwaRuntime::assetChannelCount);
+        newItem->refreshFileProperties();
         currentState->addAsset(newItem);
         currentState->setLastTouchedAsset(newItem);
         emit sendCurrentState(currentState);
-        //setCurrentState(currentState);
     }
 }
 
 void RwaStateView::deleteAssetItem(const QString &path)
 {
+    if(backend->isSimulationRunning()) // the runtime's activeAssets would keep a pointer to the deleted asset
+    {
+        qWarning() << "Cannot delete an asset while the simulation is running.";
+        return;
+    }
+
     if(currentState)
     {
         RwaAsset1 *item = currentState->getAsset(path.toStdString());
+        if(!item)
+            return;
+
         QFile file(QString::fromStdString(item->getFullPath()));
 
         if(backend->trashAsset && !backend->fileUsedByAnotherAsset(item))
         {
-            if(file.exists())
-                file.remove();
+            // the undo history can resurrect the asset entry, so the file goes
+            // to the session trash, where an undo restore finds it again; only
+            // if that fails, to the system trash (a manual recovery path)
+            if(file.exists() && !backend->moveAsset2SessionTrash(QString::fromStdString(item->getFullPath())))
+            {
+                if(!file.moveToTrash())
+                    file.remove();
+            }
         }
 
         currentState->deleteAsset(path.toStdString());
         emit sendCurrentState(currentState);
+        emit sendWriteUndo("Delete Asset");
     }
 }
-
-
-
-
-
